requirements.txt
django
requests
loguru
pyTelegramBotApi
BeautifulSoup4
psycopg2
Pillow
django-resized